Cleanup and monthly top members and hiscores job

+++ WiseOldManConnector/Models/API/Responses/WOMDelta.cs
@@ -6,6 +6,6 @@ namespace WiseOldManConnector.Models.API.Responses {
         public int End { get; set; }
 
         [JsonProperty("gained")]
-        public int Gained { get; set; }
+        public double Gained { get; set; }
     }
 }
